third_party/blink/web_tests/fast/frames/iframe-name-and-id-expected.txt

Tests: getting and setting frames[x].name on an iFrame

Related bug: rdar://problem/4170960 https://bugs.webkit.org/show_bug.cgi?id=3952 iFrame name and id information incorrectly returned

frames[0].name (should be frame0_name): frame0_name

frames[0].id (should be undefined): undefined

frames[0].name (should be frame0_newname): frame0_newname

frames[0].id (should be frame0_newid): frame0_newid
